requirements.txt
paho-mqtt
justpy
asyncio
psycopg2
